requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
cftime==1.6.4.post1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
dask==2024.11.1
decorator==5.1.1
distributed==2024.11.1
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.25.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jsonargparse==4.34.1
kiwisolver==1.4.8
lightning-utilities==0.11.7
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
ncdump-rich==0.5.2
netCDF4==1.7.2
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-click==1.8.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scipy==1.14.1
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
stack-data==0.6.3
sympy==1.13.3
tblib==3.0.0
tenacity==9.0.0
timm==1.0.9
toolz==1.0.0
torch==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.0
wcwidth==0.2.13
xarray==2024.10.0
yarl==1.12.1
zict==3.0.0
zipp==3.20.2
